requirements.txt
passlib
python-dotenv
fastapi
python-jose[cryptography]
pydantic[email]
bcrypt
streamlit
python-multipart
langchain
langchain-google-genai
chromadb
unstructured
pypdf
langchain-community
langchain-chroma
sqlalchemy[sqlite]
streamlit-browser-session-storage